feat(labs): created tags service to work with tags; updated local docker compose to have healthchecks; updated `labs_service.proto` to have proper naming of tag list;  created `tags_service.proto`; splitted code to be separate services in separate `.py` files; fixed models attributes retrieval; created tag model
