+++ services/labs-service/app/proto/labs_service.proto
@@ -33,7 +33,7 @@ message Lab {
   int64 stars_total = 9;
   int64 people_rated = 10;
   ArticleList related_articles = 11;
-  TagList tags = 12;
+  LabTagList tags = 12;
 }
 
 message ArticleList {
@@ -41,9 +41,9 @@ message ArticleList {
   repeated int64 article_id = 2;
 }
 
-message TagList {
-  int64 total_count = 1;
-  repeated string tags = 2;
+message LabTagList {
+  int32 total_count = 1;
+  repeated int32 tag_ids = 2;
 }
 
 message CreateLabRequest {
@@ -51,7 +51,7 @@ message CreateLabRequest {
   string title = 2;
   string abstract = 3;
   optional ArticleList related_articles = 4;
-  optional TagList tags = 5;
+  optional LabTagList tags = 5;
 }
 
 message GetLabRequest {
@@ -73,7 +73,7 @@ message UpdateLabRequest {
   optional string title = 2;
   optional string abstract = 3;
   optional ArticleList related_articles = 4;
-  optional TagList tags = 5;
+  optional LabTagList tags = 5;
 }
 
 message DeleteLabRequest {
